[sync experiments with arXiv v2]
