+++ experiments/outputs/adaptiveness_table.tex
@@ -4,7 +4,7 @@
 \begin{tabular}{lcccccccccc} 
 \toprule
          & \multicolumn{2}{c}{$\lambda={0}$}     & \multicolumn{2}{c}{$\lambda={0.001}$}     & \multicolumn{2}{c}{$\lambda={0.01}$}     & \multicolumn{2}{c}{$\lambda={0.1}$}     & \multicolumn{2}{c}{$\lambda={1}$}    \\ 
-         \cmidrule(r){1-2}     \cmidrule(r){3-4}     \cmidrule(r){5-6}     \cmidrule(r){7-8}     \cmidrule(r){9-10}    
+         \cmidrule(r){2-3}     \cmidrule(r){4-5}     \cmidrule(r){6-7}     \cmidrule(r){8-9}     \cmidrule(r){10-11}    
 size &cnt & cvg     &cnt & cvg     &cnt & cvg     &cnt & cvg     &cnt & cvg     \\ 
 \midrule 
 0 to 1      & 11627 & 0.88  & 11539 & 0.88  & 11225 & 0.89  & 10476 & 0.92  & 10027 & 0.93 \\
