--- experiments/outputs/exp_table1.tex
@@ -1,23 +1,0 @@
-\begin{table}[t] 
-\centering 
-\tiny 
-\begin{tabular}{lcccccccccc} 
-\toprule 
- & \multicolumn{2}{c}{Accuracy}  & \multicolumn{4}{c}{\ctwo{Coverage}{$\alpha$=0.10}} & \multicolumn{4}{c}{\ctwo{Size}{$\alpha$=0.10}} \\ 
-\cmidrule(r){2-3}  \cmidrule(r){4-7}  \cmidrule(r){8-11} \\ 
-Model & Top-1 & Top-5 & Top K & Naive & APS & RAPS & Top K & Naive & APS & RAPS \\ 
-\midrule \\ 
- ResNeXt101 &  0.793 &  0.945 & 0.9 & 0.889 & 0.9 & 0.9 & 2.421 & 17.26 & 19.719 & 2.073 \\ 
- ResNet152 &  0.783 &  0.94 & 0.9 & 0.895 & 0.9 & 0.9 & 2.629 & 9.708 & 10.367 & 2.143 \\ 
- ResNet101 &  0.774 &  0.936 & 0.9 & 0.895 & 0.9 & 0.899 & 2.835 & 10.251 & 10.843 & 2.241 \\ 
- ResNet50 &  0.762 &  0.929 & 0.9 & 0.895 & 0.9 & 0.9 & 3.137 & 11.462 & 12.21 & 2.599 \\ 
- ResNet18 &  0.698 &  0.891 & 0.9 & 0.895 & 0.9 & 0.9 & 5.748 & 15.25 & 16.101 & 4.397 \\ 
- DenseNet161 &  0.772 &  0.936 & 0.9 & 0.895 & 0.9 & 0.899 & 2.847 & 11.098 & 12.028 & 2.291 \\ 
- VGG16 &  0.716 &  0.904 & 0.9 & 0.895 & 0.9 & 0.9 & 4.749 & 13.473 & 14.019 & 3.531 \\ 
- Inception &  0.695 &  0.886 & 0.899 & 0.885 & 0.9 & 0.9 & 6.267 & 75.242 & 88.54 & 5.284 \\ 
- ShuffleNet &  0.694 &  0.883 & 0.9 & 0.891 & 0.901 & 0.9 & 6.445 & 28.5 & 32.409 & 4.989 \\ 
-\bottomrule 
-\end{tabular} 
-\caption{\textbf{Results on Imagenet-Val.} We report coverage and size of the optimal, randomized fixed sets, \naive, \aps,\ and \raps\ sets for nine different Imagenet classifiers. The median-of-means for each column is reported over 100 different trials. See Section~\ref{subsec:imagenet-val} for full details.} 
-\label{table:imagenet-val} 
-\end{table}
