+++ experiments/outputs/imagenetresults_0_1.tex
@@ -3,21 +3,21 @@
 \small 
 \begin{tabular}{lcccccccccc} 
 \toprule 
- & \multicolumn{2}{c}{Accuracy}  & \multicolumn{4}{c}{\ctwo{Coverage}{$\alpha$=0.10}} & \multicolumn{4}{c}{\ctwo{Size}{$\alpha$=0.10}} \\ 
-\cmidrule(r){2-3}  \cmidrule(r){4-7}  \cmidrule(r){8-11} \\ 
+ & \multicolumn{2}{c}{Accuracy}  & \multicolumn{4}{c}{Coverage} & \multicolumn{4}{c}{Size} \\ 
+\cmidrule(r){2-3}  \cmidrule(r){4-7}  \cmidrule(r){8-11} 
 Model & Top-1 & Top-5 & Top K & Naive & APS & RAPS & Top K & Naive & APS & RAPS \\ 
-\midrule \\ 
- ResNeXt101 &  0.793 &  0.945 & 0.9 & 0.889 & 0.9 & 0.9 & 2.42 & 17.1 & 19.7 & 2.0 \\ 
- ResNet152 &  0.783 &  0.94 & 0.9 & 0.895 & 0.9 & 0.9 & 2.63 & 9.78 & 10.4 & 2.11 \\ 
- ResNet101 &  0.774 &  0.936 & 0.9 & 0.896 & 0.9 & 0.9 & 2.83 & 10.3 & 10.7 & 2.25 \\ 
- ResNet50 &  0.761 &  0.929 & 0.9 & 0.896 & 0.9 & 0.9 & 3.14 & 11.8 & 12.3 & 2.57 \\ 
- ResNet18 &  0.698 &  0.891 & 0.9 & 0.895 & 0.9 & 0.9 & 5.72 & 15.5 & 16.2 & 4.43 \\ 
- DenseNet161 &  0.771 &  0.936 & 0.9 & 0.894 & 0.9 & 0.9 & 2.84 & 11.2 & 12.1 & 2.29 \\ 
- VGG16 &  0.716 &  0.904 & 0.9 & 0.895 & 0.901 & 0.9 & 4.75 & 13.4 & 14.1 & 3.54 \\ 
- Inception &  0.695 &  0.887 & 0.9 & 0.885 & 0.9 & 0.901 & 6.3 & 75.4 & 89.1 & 5.32 \\ 
- ShuffleNet &  0.694 &  0.883 & 0.9 & 0.891 & 0.9 & 0.9 & 6.46 & 28.9 & 31.9 & 5.05 \\ 
+\midrule 
+ ResNeXt101 &  0.793 &  0.945 & 0.9 & 0.889 & 0.9 & 0.9 & 2.42 & 17.2 & 19.9 & 2.01 \\ 
+ ResNet152 &  0.783 &  0.94 & 0.9 & 0.894 & 0.9 & 0.9 & 2.64 & 9.68 & 10.4 & 2.09 \\ 
+ ResNet101 &  0.774 &  0.936 & 0.9 & 0.895 & 0.9 & 0.9 & 2.83 & 10.0 & 10.8 & 2.25 \\ 
+ ResNet50 &  0.761 &  0.929 & 0.899 & 0.896 & 0.9 & 0.9 & 3.13 & 11.7 & 12.3 & 2.55 \\ 
+ ResNet18 &  0.698 &  0.891 & 0.9 & 0.895 & 0.9 & 0.9 & 5.74 & 15.3 & 16.1 & 4.38 \\ 
+ DenseNet161 &  0.772 &  0.936 & 0.9 & 0.894 & 0.9 & 0.9 & 2.84 & 11.2 & 12.0 & 2.29 \\ 
+ VGG16 &  0.716 &  0.904 & 0.9 & 0.896 & 0.901 & 0.9 & 4.75 & 13.4 & 14.1 & 3.54 \\ 
+ Inception &  0.695 &  0.886 & 0.899 & 0.884 & 0.9 & 0.899 & 6.27 & 74.8 & 88.8 & 5.24 \\ 
+ ShuffleNet &  0.694 &  0.883 & 0.9 & 0.892 & 0.9 & 0.899 & 6.45 & 28.8 & 32.1 & 5.01 \\ 
 \bottomrule 
 \end{tabular} 
-\caption{\textbf{Results on Imagenet-Val.} We report coverage and size of the optimal, randomized fixed sets, \naive, \aps,\ and \raps\ sets for nine different Imagenet classifiers. The median-of-means for each column is reported over 100 different trials. See Section~\ref{subsec:imagenet-val} for full details.} 
+\caption{\textbf{Results on Imagenet-Val.} We report coverage and size of the optimal, randomized fixed sets, \naive, \aps,\ and \raps\ sets for nine different Imagenet classifiers. The median-of-means for each column is reported over 100 different trials at the 10\% level. See Section~\ref{subsec:imagenet-val} for full details.} 
 \label{table:imagenet-val} 
 \end{table}
