+++ experiments/outputs/imagenetv2results_0_1.tex
@@ -3,21 +3,21 @@
 \small 
 \begin{tabular}{lcccccccccc} 
 \toprule 
- & \multicolumn{2}{c}{Accuracy}  & \multicolumn{4}{c}{\ctwo{Coverage}{$\alpha$=0.10}} & \multicolumn{4}{c}{\ctwo{Size}{$\alpha$=0.10}} \\ 
-\cmidrule(r){2-3}  \cmidrule(r){4-7}  \cmidrule(r){8-11} \\ 
+ & \multicolumn{2}{c}{Accuracy}  & \multicolumn{4}{c}{Coverage} & \multicolumn{4}{c}{Size} \\ 
+\cmidrule(r){2-3}  \cmidrule(r){4-7}  \cmidrule(r){8-11} 
 Model & Top-1 & Top-5 & Top K & Naive & APS & RAPS & Top K & Naive & APS & RAPS \\ 
-\midrule \\ 
- ResNeXt101 &  0.679 &  0.875 & 0.9 & 0.888 & 0.9 & 0.9 & 7.6 & 43.2 & 51.5 & 6.43 \\ 
- ResNet152 &  0.67 &  0.876 & 0.9 & 0.895 & 0.901 & 0.9 & 7.27 & 25.6 & 27.4 & 5.84 \\ 
- ResNet101 &  0.657 &  0.86 & 0.9 & 0.893 & 0.899 & 0.9 & 9.11 & 28.2 & 30.6 & 7.03 \\ 
- ResNet50 &  0.634 &  0.846 & 0.901 & 0.894 & 0.9 & 0.901 & 10.5 & 30.4 & 32.7 & 8.06 \\ 
- ResNet18 &  0.572 &  0.802 & 0.899 & 0.894 & 0.901 & 0.9 & 17.2 & 35.5 & 37.2 & 13.8 \\ 
- DenseNet161 &  0.652 &  0.861 & 0.899 & 0.894 & 0.899 & 0.9 & 8.41 & 29.8 & 32.1 & 7.04 \\ 
- VGG16 &  0.588 &  0.816 & 0.898 & 0.897 & 0.9 & 0.899 & 14.9 & 32.0 & 32.9 & 11.5 \\ 
- Inception &  0.573 &  0.798 & 0.9 & 0.892 & 0.899 & 0.9 & 21.7 & 143.0 & 153.0 & 21.0 \\ 
- ShuffleNet &  0.56 &  0.781 & 0.901 & 0.892 & 0.9 & 0.9 & 26.3 & 66.4 & 71.4 & 23.4 \\ 
+\midrule 
+ ResNeXt101 &  0.678 &  0.874 & 0.9 & 0.888 & 0.899 & 0.899 & 7.48 & 43.0 & 50.8 & 6.18 \\ 
+ ResNet152 &  0.67 &  0.876 & 0.899 & 0.896 & 0.9 & 0.9 & 7.18 & 25.8 & 27.2 & 5.69 \\ 
+ ResNet101 &  0.657 &  0.859 & 0.901 & 0.894 & 0.9 & 0.898 & 9.21 & 28.7 & 30.7 & 6.93 \\ 
+ ResNet50 &  0.634 &  0.847 & 0.898 & 0.894 & 0.899 & 0.9 & 10.3 & 30.3 & 32.3 & 7.8 \\ 
+ ResNet18 &  0.572 &  0.802 & 0.902 & 0.895 & 0.9 & 0.9 & 17.5 & 35.3 & 37.4 & 13.3 \\ 
+ DenseNet161 &  0.653 &  0.862 & 0.902 & 0.895 & 0.901 & 0.901 & 8.6 & 29.9 & 32.4 & 6.93 \\ 
+ VGG16 &  0.588 &  0.817 & 0.902 & 0.897 & 0.9 & 0.899 & 15.1 & 31.9 & 32.8 & 11.2 \\ 
+ Inception &  0.573 &  0.797 & 0.9 & 0.893 & 0.9 & 0.899 & 21.8 & 145.0 & 155.0 & 20.5 \\ 
+ ShuffleNet &  0.559 &  0.781 & 0.899 & 0.892 & 0.9 & 0.899 & 26.0 & 66.2 & 71.7 & 22.5 \\ 
 \bottomrule 
 \end{tabular} 
-\caption{\textbf{Results on Imagenet-V2.} We report coverage and size of the optimal, randomized fixed sets, \naive, \aps,\ and \raps\ sets for nine different Imagenet classifiers. The median-of-means for each column is reported over 100 different trials. See Section~\ref{subsec:imagenet-v2} for full details.} 
+\caption{\textbf{Results on Imagenet-V2.} We report coverage and size of the optimal, randomized fixed sets, \naive, \aps,\ and \raps\ sets for nine different Imagenet classifiers. The median-of-means for each column is reported over 100 different trials at the 10\% level. See Section~\ref{subsec:imagenet-v2} for full details.} 
 \label{table:imagenet-v2} 
 \end{table}
