--- experiments/outputs/inv2_exp_table2.tex
@@ -1,60 +1,0 @@
-\begin{tabular}{lllrrrrr}
-\toprule
-{} &        Model & Predictor &   Top1 &   Top5 &  alpha &  Coverage &     Size \\
-\midrule
-0  &   ResNeXt101 &     Naive &  0.679 &  0.875 &   0.05 &     0.937 &   85.350 \\
-1  &   ResNeXt101 &       APS &  0.679 &  0.875 &   0.05 &     0.949 &  105.851 \\
-2  &   ResNeXt101 &      RAPS &  0.678 &  0.875 &   0.05 &     0.950 &   21.236 \\
-3  &   ResNeXt101 &     Naive &  0.678 &  0.875 &   0.10 &     0.887 &   42.650 \\
-4  &   ResNeXt101 &       APS &  0.680 &  0.875 &   0.10 &     0.901 &   51.337 \\
-5  &   ResNeXt101 &      RAPS &  0.679 &  0.875 &   0.10 &     0.900 &    6.769 \\
-6  &    ResNet152 &     Naive &  0.671 &  0.876 &   0.05 &     0.943 &   50.496 \\
-7  &    ResNet152 &       APS &  0.670 &  0.875 &   0.05 &     0.950 &   57.369 \\
-8  &    ResNet152 &      RAPS &  0.671 &  0.875 &   0.05 &     0.950 &   20.361 \\
-9  &    ResNet152 &     Naive &  0.671 &  0.876 &   0.10 &     0.896 &   25.933 \\
-10 &    ResNet152 &       APS &  0.670 &  0.876 &   0.10 &     0.901 &   27.450 \\
-11 &    ResNet152 &      RAPS &  0.671 &  0.876 &   0.10 &     0.900 &    6.616 \\
-12 &    ResNet101 &     Naive &  0.656 &  0.859 &   0.05 &     0.943 &   55.125 \\
-13 &    ResNet101 &       APS &  0.657 &  0.860 &   0.05 &     0.950 &   62.260 \\
-14 &    ResNet101 &      RAPS &  0.656 &  0.860 &   0.05 &     0.949 &   24.613 \\
-15 &    ResNet101 &     Naive &  0.657 &  0.861 &   0.10 &     0.894 &   28.306 \\
-16 &    ResNet101 &       APS &  0.657 &  0.860 &   0.10 &     0.900 &   30.342 \\
-17 &    ResNet101 &      RAPS &  0.657 &  0.859 &   0.10 &     0.899 &    8.441 \\
-18 &     ResNet50 &     Naive &  0.634 &  0.847 &   0.05 &     0.944 &   59.076 \\
-19 &     ResNet50 &       APS &  0.632 &  0.846 &   0.05 &     0.950 &   65.720 \\
-20 &     ResNet50 &      RAPS &  0.634 &  0.846 &   0.05 &     0.950 &   29.236 \\
-21 &     ResNet50 &     Naive &  0.633 &  0.847 &   0.10 &     0.894 &   30.463 \\
-22 &     ResNet50 &       APS &  0.633 &  0.846 &   0.10 &     0.900 &   32.580 \\
-23 &     ResNet50 &      RAPS &  0.634 &  0.847 &   0.10 &     0.901 &    9.636 \\
-24 &     ResNet18 &     Naive &  0.571 &  0.802 &   0.05 &     0.942 &   64.868 \\
-25 &     ResNet18 &       APS &  0.572 &  0.801 &   0.05 &     0.950 &   73.757 \\
-26 &     ResNet18 &      RAPS &  0.570 &  0.801 &   0.05 &     0.950 &   48.192 \\
-27 &     ResNet18 &     Naive &  0.572 &  0.802 &   0.10 &     0.894 &   35.099 \\
-28 &     ResNet18 &       APS &  0.572 &  0.803 &   0.10 &     0.901 &   37.713 \\
-29 &     ResNet18 &      RAPS &  0.572 &  0.802 &   0.10 &     0.901 &   16.600 \\
-30 &  DenseNet161 &     Naive &  0.653 &  0.861 &   0.05 &     0.941 &   59.740 \\
-31 &  DenseNet161 &       APS &  0.654 &  0.862 &   0.05 &     0.950 &   72.477 \\
-32 &  DenseNet161 &      RAPS &  0.654 &  0.861 &   0.05 &     0.950 &   25.231 \\
-33 &  DenseNet161 &     Naive &  0.653 &  0.861 &   0.10 &     0.894 &   29.398 \\
-34 &  DenseNet161 &       APS &  0.653 &  0.861 &   0.10 &     0.901 &   31.919 \\
-35 &  DenseNet161 &      RAPS &  0.652 &  0.862 &   0.10 &     0.900 &    7.680 \\
-36 &        VGG16 &     Naive &  0.588 &  0.816 &   0.05 &     0.943 &   57.361 \\
-37 &        VGG16 &       APS &  0.588 &  0.816 &   0.05 &     0.951 &   65.077 \\
-38 &        VGG16 &      RAPS &  0.588 &  0.816 &   0.05 &     0.950 &   38.188 \\
-39 &        VGG16 &     Naive &  0.589 &  0.816 &   0.10 &     0.897 &   31.702 \\
-40 &        VGG16 &       APS &  0.588 &  0.816 &   0.10 &     0.900 &   32.829 \\
-41 &        VGG16 &      RAPS &  0.589 &  0.816 &   0.10 &     0.899 &   14.400 \\
-42 &    Inception &     Naive &  0.574 &  0.798 &   0.05 &     0.943 &  250.736 \\
-43 &    Inception &       APS &  0.573 &  0.797 &   0.05 &     0.950 &  275.085 \\
-44 &    Inception &      RAPS &  0.572 &  0.797 &   0.05 &     0.950 &   72.424 \\
-45 &    Inception &     Naive &  0.574 &  0.797 &   0.10 &     0.893 &  144.867 \\
-46 &    Inception &       APS &  0.573 &  0.797 &   0.10 &     0.900 &  155.394 \\
-47 &    Inception &      RAPS &  0.573 &  0.797 &   0.10 &     0.900 &   20.715 \\
-48 &   ShuffleNet &     Naive &  0.559 &  0.781 &   0.05 &     0.942 &  125.593 \\
-49 &   ShuffleNet &       APS &  0.560 &  0.782 &   0.05 &     0.950 &  144.303 \\
-50 &   ShuffleNet &      RAPS &  0.560 &  0.781 &   0.05 &     0.951 &   80.184 \\
-51 &   ShuffleNet &     Naive &  0.559 &  0.781 &   0.10 &     0.893 &   66.232 \\
-52 &   ShuffleNet &       APS &  0.560 &  0.781 &   0.10 &     0.900 &   71.756 \\
-53 &   ShuffleNet &      RAPS &  0.559 &  0.780 &   0.10 &     0.900 &   25.534 \\
-\bottomrule
-\end{tabular}
